+++ crates/https/Cargo.toml
@@ -59,8 +59,7 @@ thiserror = "1.0.20"
 tokio = { version = "1.0", features = ["io-util", "net", "rt"] }
 tokio-rustls = "0.22.0"
 # disables default features, i.e. openssl...
-trust-dns-proto = { version = "0.20.1", path = "../proto", features = ["tokio-runtime"], default-features = false }
-trust-dns-rustls = { version = "0.20.1", path = "../rustls", default-features = false }
+trust-dns-proto = { version = "0.20.1", path = "../proto", features = ["tokio-runtime", "dns-over-rustls"], default-features = false }
 webpki-roots = "0.21"
 webpki = "0.21"
