+++ crates/proto/Cargo.toml
@@ -36,10 +36,15 @@ codecov = { repository = "bluejekyll/trust-dns", branch = "main", service = "git
 maintenance = { status = "actively-developed" }
 
 [features]
+dns-over-tls = []
+dns-over-rustls = ["dns-over-tls", "rustls", "tokio-rustls", "webpki"]
+
 dnssec-openssl = ["dnssec", "openssl"]
 dnssec-ring = ["dnssec", "ring"]
 dnssec = []
+
 testing = []
+
 tokio-runtime = ["tokio/net", "tokio/rt", "tokio/time", "tokio/rt-multi-thread"]
 default = ["tokio-runtime"]
 
@@ -74,18 +79,22 @@ log = "0.4"
 openssl = { version = "0.10", features = ["v102", "v110"], optional = true }
 rand = "0.8"
 ring = { version = "0.16", optional = true, features = ["std"] }
+rustls = { version = "0.19", optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
 smallvec = "1.6"
 socket2 = { version = "0.4.0", optional = true }
 thiserror = "1.0.20"
 tinyvec = { version = "1.1.1", features = ["alloc"] }
 tokio = { version = "1.0", optional = true }
+tokio-rustls = { version = "0.22", optional = true, features = ["early-data"] }
 url = "2.1.0"
 wasm-bindgen-crate = { version = "0.2.58", optional = true, package = "wasm-bindgen" }
+webpki = { version = "0.21", optional = true }
 
 [dev-dependencies]
 env_logger = "0.8"
 futures-executor = { version = "0.3.5", default-features = false, features = ["std"] }
+#openssl = { version = "0.10", features = ["v102", "v110"] }
 tokio = { version = "1.0", features = ["rt", "time"] }
 
 [package.metadata.docs.rs]
