+++ crates/proto/src/rustls/tests.rs
@@ -29,10 +29,9 @@ use rustls::ClientConfig;
 use tokio::net::TcpStream as TokioTcpStream;
 use tokio::runtime::Runtime;
 
-use trust_dns_proto::xfer::SerialMessage;
-use trust_dns_proto::{iocompat::AsyncIoTokioAsStd, DnsStreamHandle};
-
-use crate::tls_connect;
+use crate::rustls::tls_connect;
+use crate::xfer::SerialMessage;
+use crate::{iocompat::AsyncIoTokioAsStd, DnsStreamHandle};
 
 // this fails on linux for some reason. It appears that a buffer somewhere is dirty
 //  and subsequent reads of a message buffer reads the wrong length. It works for 2 iterations
