+++ tests/integration-tests/Cargo.toml
@@ -63,7 +63,7 @@ dns-over-https = []
 # TODO: need to make server support rustls and native-tls
 # dns-over-native-tls = ["dns-over-tls",  "trust-dns-resolver/dns-over-native-tls", "trust-dns-server/dns-over-native-tls"]
 dns-over-openssl = ["dns-over-tls", "trust-dns-resolver/dns-over-openssl", "trust-dns-server/dns-over-openssl"]
-dns-over-rustls = ["dns-over-tls", "trust-dns-resolver/dns-over-rustls", "trust-dns-server/dns-over-rustls"]
+dns-over-rustls = ["dns-over-tls", "trust-dns-proto/dns-over-rustls", "trust-dns-resolver/dns-over-rustls", "trust-dns-server/dns-over-rustls"]
 dns-over-tls = []
 
 sqlite = ["trust-dns-server/sqlite"]
@@ -84,7 +84,6 @@ trust-dns-https = { version = "0.20.1", path = "../../crates/https" }
 trust-dns-openssl = { version = "0.20.1", path = "../../crates/openssl" }
 trust-dns-proto = { version = "0.20.1", path = "../../crates/proto", features = ["testing"] }
 trust-dns-resolver = { version = "0.20.1", path = "../../crates/resolver" }
-trust-dns-rustls = { version = "0.20.1", path = "../../crates/rustls" }
 # TODO: fixup tests to not require openssl
 trust-dns-server = { version = "0.20.1", path = "../../crates/server" }
 webpki-roots = { version = "0.21", optional = true }
